+++ src/event/impls.rs
@@ -1,7 +1,7 @@
 use serde_json::json;
 
-use super::*;
 use crate::error::{Cause, Error};
+use crate::prelude::*;
 use crate::protocol::ProtocolEvent;
 use crate::utils;
